Clean up search result component
